__app__/requirements.txt
azure-functions
azure-keyvault
azure-identity
azure-storage-blob
beautifulsoup4
django
requests